Finished Refactoring DrawState.kt Added factory methods to DynamicColor.kt

+++ mechanica/src/main/kotlin/com/mechanica/engine/drawer/subclasses/layout/OriginDrawerImpl.kt
@@ -6,7 +6,7 @@ import com.mechanica.engine.drawer.state.DrawState
 class OriginDrawerImpl(drawer: Drawer, private val state: DrawState) : OriginDrawer, Drawer by drawer {
 
     override fun normalized(x: Number, y: Number): Drawer {
-        state.origin.normalized.set(x, y)
+        state.origin.normalized.set(x.toDouble(), y.toDouble())
         return this
     }
